MAINT-1171 Admin function to restore effectiveTime on content after bad merge.
Occasionally a concept conflict does not resolve correctly and the effectiveTime values are not retained on the merged content. The RVF validation report catches this issue. The bug in the terminology server will be fixed once we are able to reproduce the issue.
This admin function will automatically restore the effectiveTime on a list of given components, where appropriate, to make fixing the content quick and easy until the code fix is written.